Moved the little (mostly worthless?) test app for SoundFieldControl from its hiding place in SIL.Media.Tests to a folder under TestApps, enabling us to remove dependency on SIL.Windows.Forms from SIL.Media.Tests
