+++ SIL.Media.Tests/SIL.Media.Tests.csproj
@@ -27,6 +27,5 @@
     <ProjectReference Include="..\SIL.Core\SIL.Core.csproj" />
     <ProjectReference Include="..\SIL.Media\SIL.Media.csproj" />
     <ProjectReference Include="..\SIL.TestUtilities\SIL.TestUtilities.csproj" />
-    <ProjectReference Include="..\SIL.Windows.Forms\SIL.Windows.Forms.csproj" />
   </ItemGroup>
 </Project>
