+++ TestApps/SoundFieldControlTestApp/MainForm.cs
@@ -1,12 +1,12 @@
-﻿using System;
+using System;
 using System.IO;
 using System.Windows.Forms;
 
-namespace SIL.Media.Tests
+namespace SIL.Media.SoundFieldControlTestApp
 {
-	public partial class Form1 : Form
+	public partial class MainForm : Form
 	{
-		public Form1()
+		public MainForm()
 		{
 //            _recorder = new AudioIrrKlangSession(_path);
 			InitializeComponent();
@@ -106,23 +106,9 @@ namespace SIL.Media.Tests
 //            UpdateScreen();
 //        }
 //
-//        private void Form1_Load(object sender, EventArgs e)
+//        private void MainForm_Load(object sender, EventArgs e)
 //        {
 //           // _recorder.Test();
 //        }
-	}
-
-	static class Program
-	{
-		/// <summary>
-		/// The main entry point for the application.
-		/// </summary>
-		[STAThread]
-		static void Main()
-		{
-			Application.EnableVisualStyles();
-			Application.SetCompatibleTextRenderingDefault(false);
-			Application.Run(new Form1());
-		}
 	}
 }
