+++ TestApps/SoundFieldControlTestApp/Program.cs
@@ -1,0 +1,19 @@
+using System;
+using System.Windows.Forms;
+
+namespace SIL.Media.SoundFieldControlTestApp
+{
+	internal static class Program
+	{
+		/// <summary>
+		/// The main entry point for the application.
+		/// </summary>
+		[STAThread]
+		static void Main()
+		{
+			Application.EnableVisualStyles();
+			Application.SetCompatibleTextRenderingDefault(false);
+			Application.Run(new MainForm());
+		}
+	}
+}
